+++ src/morphic.js
@@ -491,6 +491,19 @@
     currently pressed mouse button, which is either 'left' or 'right'.
     You can use this to let users interact with 3D environments.
 
+    Mouse move and click events are only triggered when and as long as the
+    mouse pointer is within a morph's bounds. If you wish to confine and
+    retain mouse interaction to a particular morph while a mouse button is
+    pressed you can call the morph's
+    
+        lockMouseFocus()
+
+    method, preferrably in one of its mouseDown methods. Afterwards the
+    ensuing mouse events will be routed to the locked morph even after the
+    mouse pointer has left its bounds, and the mouseLeave event will not
+    occur until the mouse button is released again. This is useful for
+    creating sliders and handle widgets for resizing, moving, rotating etc.
+
     The
 
         mouseEnterDragging(morph)
@@ -1338,7 +1351,7 @@
 
 /*jshint esversion: 11, bitwise: false*/
 
-var morphicVersion = '2025-March-17';
+var morphicVersion = '2025-March-19';
 var modules = {}; // keep track of additional loaded modules
 var useBlurredShadows = true;
 
@@ -4766,6 +4779,16 @@ Morph.prototype.escalateEvent = function (functionName, arg) {
     }
 };
 
+Morph.prototype.lockMouseFocus = function () {
+    // while the mouse button is pressed let me handle mouseMove events
+    // and the next mouseUp event, and don't fire the mouseLeave event.
+    // Use for widgets that are dragged around, e.g. resize handles,
+    // dials, sliders etc.
+    var hand = this.world().hand;
+    hand.inputTarget = this;
+    hand.clickTarget = this;
+};
+
 // Morph eval:
 
 Morph.prototype.evaluateString = function (code) {
@@ -4877,6 +4900,7 @@ HandleMorph.prototype.init = function (
     this.target = target || null;
     this.minExtent = new Point(minX || 0, minY || 0);
     this.inset = new Point(insetX || 0, insetY || insetX || 0);
+    this.offset = null;
     this.type =  type || 'resize'; // also: 'move', 'moveCenter', 'movePivot'
     this.isHighlighted = false;
     HandleMorph.uber.init.call(this);
@@ -5046,60 +5070,6 @@ HandleMorph.prototype.renderHandleOn = function (
     }
 };
 
-// HandleMorph stepping:
-
-HandleMorph.prototype.step = null;
-
-HandleMorph.prototype.mouseDownLeft = function (pos) {
-    var world = this.root(),
-        offset;
-
-    if (!this.target) {
-        return null;
-    }
-    if (this.type.indexOf('move') === 0) {
-        offset = pos.subtract(this.center());
-    } else {
-        offset = pos.subtract(this.bounds.origin);
-    }
-
-    this.step = () => {
-        var newPos, newExt;
-        if (world.hand.mouseButton) {
-            newPos = world.hand.bounds.origin.copy().subtract(offset);
-            if (this.type === 'resize') {
-                newExt = newPos.add(
-                    this.extent().add(this.inset)
-                ).subtract(this.target.bounds.origin);
-                newExt = newExt.max(this.minExtent);
-                this.target.setExtent(newExt);
-
-                this.setPosition(
-                    this.target.bottomRight().subtract(
-                        this.extent().add(this.inset)
-                    )
-                );
-            } else if (this.type === 'moveCenter') {
-                this.target.setCenter(newPos);
-            } else if (this.type === 'movePivot') {
-                this.target.setPivot(newPos);
-                this.setCenter(this.target.rotationCenter());
-            } else { // type === 'move'
-                this.target.setPosition(
-                    newPos.subtract(this.target.extent())
-                        .add(this.extent())
-                );
-            }
-        } else {
-            this.step = null;
-        }
-    };
-
-    if (!this.target.step) {
-        this.target.step = nop;
-    }
-};
-
 // HandleMorph dragging and dropping:
 
 HandleMorph.prototype.rootForGrab = function () {
@@ -5116,6 +5086,47 @@ HandleMorph.prototype.mouseEnter = function () {
 HandleMorph.prototype.mouseLeave = function () {
     this.isHighlighted = false;
     this.changed();
+};
+
+HandleMorph.prototype.mouseDownLeft = function (pos) {
+    if (!this.target) {
+        return null;
+    }
+    if (this.type.indexOf('move') === 0) {
+        this.offset = pos.subtract(this.center());
+    } else {
+        this.offset = pos.subtract(this.bounds.origin);
+    }
+    this.lockMouseFocus();
+};
+
+HandleMorph.prototype.mouseMove = function (pos) {
+    var newPos, newExt;
+
+    newPos = pos.subtract(this.offset);
+    if (this.type === 'resize') {
+        newExt = newPos.add(
+            this.extent().add(this.inset)
+        ).subtract(this.target.bounds.origin);
+        newExt = newExt.max(this.minExtent);
+        this.target.setExtent(newExt);
+
+        this.setPosition(
+            this.target.bottomRight().subtract(
+                this.extent().add(this.inset)
+            )
+        );
+    } else if (this.type === 'moveCenter') {
+        this.target.setCenter(newPos);
+    } else if (this.type === 'movePivot') {
+        this.target.setPivot(newPos);
+        this.setCenter(this.target.rotationCenter());
+    } else { // type === 'move'
+        this.target.setPosition(
+            newPos.subtract(this.target.extent())
+                .add(this.extent())
+        );
+    }
 };
 
 // HandleMorph menu:
@@ -6706,23 +6717,17 @@ DialMorph.prototype.render = function (ctx) {
     ctx.fill();
 };
 
-// DialMorph stepping:
-
-DialMorph.prototype.step = null;
+// DialMorph events:
 
 DialMorph.prototype.mouseDownLeft = function (pos) {
-    var world = this.root();
+    this.lockMouseFocus();
+};
 
-    this.step = () => {
-        if (world.hand.mouseButton) {
-            this.setValue(
-            	this.getValueOf(world.hand.bounds.origin),
-             	world.currentKey !== 16 // snap to tick
-            );
-        } else {
-            this.step = null;
-        }
-    };
+DialMorph.prototype.mouseMove = function (pos) {
+    this.setValue(
+        this.getValueOf(pos),
+        this.world().currentKey !== 16 // snap to tick
+    );
 };
 
 // DialMorph menu:
@@ -7053,9 +7058,9 @@ SliderButtonMorph.prototype.mouseClickLeft = function () {
     this.rerender();
 };
 
-SliderButtonMorph.prototype.mouseMove = function () {
-    // prevent my parent from getting picked up
-    nop();
+SliderButtonMorph.prototype.mouseMove = function (pos) {
+    // delegate event to the slider
+    this.escalateEvent('mouseMove', pos);
 };
 
 // SliderMorph ///////////////////////////////////////////////////
@@ -7359,49 +7364,43 @@ SliderMorph.prototype.numericalSetters = function () {
     return list;
 };
 
-// SliderMorph stepping:
-
-SliderMorph.prototype.step = null;
+// SliderMorph events:
 
 SliderMorph.prototype.mouseDownLeft = function (pos) {
-    var world;
-
-    if (!this.button.bounds.containsPoint(pos)) {
-        this.offset = new Point(); // return null;
-    } else {
+    if (this.button.bounds.containsPoint(pos)) {
         this.offset = pos.subtract(this.button.bounds.origin);
+    } else {
+        this.offset = new Point();
+        this.button.userState = 'pressed';
+        this.button.rerender();
     }
-    world = this.root();
+    this.button.lockMouseFocus();
+    this.mouseMove(pos);
+};
 
-    this.step = () => {
-        var mousePos, newX, newY;
-        if (world.hand.mouseButton) {
-            mousePos = world.hand.bounds.origin;
-            if (this.orientation === 'vertical') {
-                newX = this.button.bounds.origin.x;
-                newY = Math.max(
-                    Math.min(
-                        mousePos.y - this.offset.y,
-                        this.bottom() - this.button.height()
-                    ),
-                    this.top()
-                );
-            } else {
-                newY = this.button.bounds.origin.y;
-                newX = Math.max(
-                    Math.min(
-                        mousePos.x - this.offset.x,
-                        this.right() - this.button.width()
-                    ),
-                    this.left()
-                );
-            }
-            this.button.setPosition(new Point(newX, newY));
-            this.updateValue();
-        } else {
-            this.step = null;
-        }
-    };
+SliderMorph.prototype.mouseMove = function (pos) {
+    var newX, newY;
+    if (this.orientation === 'vertical') {
+        newX = this.button.bounds.origin.x;
+        newY = Math.max(
+            Math.min(
+                pos.y - this.offset.y,
+                this.bottom() - this.button.height()
+            ),
+            this.top()
+        );
+    } else {
+        newY = this.button.bounds.origin.y;
+        newX = Math.max(
+            Math.min(
+                pos.x - this.offset.x,
+                this.right() - this.button.width()
+            ),
+            this.left()
+        );
+    }
+    this.button.setPosition(new Point(newX, newY));
+    this.updateValue();
 };
 
 // MouseSensorMorph ////////////////////////////////////////////////////
@@ -11197,6 +11196,8 @@ HandMorph.prototype.init = function (aWorld) {
     this.mouseOverList = [];
     this.mouseOverBounds = [];
     this.morphToGrab = null;
+    this.inputTarget = null;
+    this.clickTarget = null;
     this.grabPosition = null;
     this.grabOrigin = null;
     this.temporaries = [];
@@ -11405,8 +11406,11 @@ HandMorph.prototype.processMouseDown = function (event) {
     if (this.children.length !== 0) {
         this.drop();
         this.mouseButton = null;
+        this.inputTarget = null;
+        this.clickTarget = null;
     } else {
         morph = this.morphAtPointer();
+        this.clickTarget = morph;
         if (this.world.activeMenu) {
             if (!contains(
                     morph.allParents(),
@@ -11492,11 +11496,10 @@ HandMorph.prototype.processTouchEnd = function (event) {
 };
 
 HandMorph.prototype.processMouseUp = function () {
-    var morph = this.morphAtPointer(),
+    var morph = this.inputTarget || this.morphAtPointer(),
         context,
         contextMenu,
         expectedClick;
-
     this.destroyTemporaries();
     if (this.children.length !== 0) {
         this.drop();
@@ -11521,9 +11524,19 @@ HandMorph.prototype.processMouseUp = function () {
         while (!morph[expectedClick]) {
             morph = morph.parent;
         }
-        morph[expectedClick](this.bounds.origin);
+        if (this.clickTarget && this.clickTarget.allParents().includes(morph)) {
+            morph[expectedClick](this.bounds.origin);
+            if (this.inputTarget &&
+                !this.inputTarget.bounds.containsPoint(this.bounds.origin) &&
+                this.inputTarget.mouseLeave
+            ) {
+                this.inputTarget.mouseLeave();
+            }
+        }
     }
     this.mouseButton = null;
+    this.inputTarget = null;
+    this.clickTarget = null;
 };
 
 HandMorph.prototype.processDoubleClick = function () {
@@ -11559,7 +11572,9 @@ HandMorph.prototype.processMouseMove = function (event) {
     this.setPosition(pos);
 
     // determine the new mouse-over-list:
-    mouseOverNew = this.morphAtPointer().allParents();
+    mouseOverNew = this.morphAtPointer().allParents().concat(
+        this.inputTarget ? this.inputTarget.allParents() : []
+    );
     mouseOverBoundsNew = mouseOverNew.filter(m => m.isVisible &&
         m.visibleBounds().containsPoint(this.bounds.origin) &&
             !m.holes.some(any =>
@@ -11567,9 +11582,12 @@ HandMorph.prototype.processMouseMove = function (event) {
     );
 
     if (!this.children.length && this.mouseButton) {
-        topMorph = this.morphAtPointer();
+        topMorph = this.inputTarget || this.morphAtPointer();
         morph = topMorph.rootForGrab();
-        if (topMorph.mouseMove) {
+        if (this.clickTarget &&
+            this.clickTarget.allParents().includes(topMorph) &&
+            topMorph.mouseMove
+        ) {
             topMorph.mouseMove(pos, this.mouseButton);
             if (this.mouseButton === 'right') {
                 this.contextMenuEnabled = false;