+++ sw.js
@@ -1,6 +1,6 @@
 /*global self, caches*/
 /*jshint esversion: 6*/
-var snapVersion = '10.6.1-250319-dev',
+var snapVersion = '10.6.1',
     cacheName = `snap-pwa-${snapVersion}`,
     filesToCache = [
         'snap.html',