+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## in development:
 
 ### 2025-03-17
+* new dev version
 * morphic: simplified BoxMorph & CircleBoxMorph outlinePath rendering with new Canvas roundRect() primitive
 * objects, widgets: simplified DialogBoxMorph and StagePicker outlinePath rendering with new Canvas roundRect primitive